requirements.txt
discord.py
python-dotenv
urlextract
Unalix
